Refactor: domain 폴더 생성 후 util 파일 재분리
